requirements.txt
assemblyai>=0.17.0
python-dotenv>=1.0.0 